Issue #SB-26318 feat: Moved the Slug class to jobs-code module and removed from asset-enrichment, auto-creator-v2 and publish-core module.
