+++ publish-pipeline/publish-core/pom.xml
@@ -37,11 +37,6 @@
             </exclusions>
         </dependency>
         <dependency>
-            <groupId>com.moparisthebest</groupId>
-            <artifactId>junidecode</artifactId>
-            <version>0.1.1</version>
-        </dependency>
-        <dependency>
             <groupId>org.scalatest</groupId>
             <artifactId>scalatest_2.12</artifactId>
             <version>3.0.6</version>
