+++ publish-pipeline/publish-core/src/main/scala/org/sunbird/job/publish/helpers/QuestionPdfGenerator.scala
@@ -2,10 +2,10 @@ package org.sunbird.job.publish.helpers
 
 import com.google.gson.Gson
 import org.slf4j.LoggerFactory
-import org.sunbird.job.publish.core.{ObjectData, Slug}
+import org.sunbird.job.publish.core.ObjectData
 import org.sunbird.job.publish.handler.{QuestionHandlerFactory, QuestionTypeHandler}
 import org.sunbird.job.publish.util.{CloudStorageUtil, FileUtils}
-import org.sunbird.job.util.HttpUtil
+import org.sunbird.job.util.{HttpUtil, Slug}
 
 import java.io.{BufferedWriter, File, FileWriter}
