+++ publish-pipeline/publish-core/src/main/scala/org/sunbird/job/publish/helpers/ThumbnailGenerator.scala
@@ -4,12 +4,13 @@ import org.apache.commons.io.FilenameUtils
 import org.apache.commons.lang3.StringUtils
 import org.imgscalr.Scalr
 import org.slf4j.LoggerFactory
-import org.sunbird.job.publish.core.{ObjectData, Slug}
+import org.sunbird.job.publish.core.ObjectData
 import org.sunbird.job.publish.util.{CloudStorageUtil, FileUtils}
+import org.sunbird.job.util.Slug
+
 import java.awt.image.BufferedImage
 import java.io.File
 import java.net.{HttpURLConnection, URL}
-
 import javax.imageio.ImageIO
 
 trait ThumbnailGenerator {
